Fix html content structure
The html file now has three main divs of information. The hero
and footer section remains the same.
Now the random information, quote section, and the sign up section
are all inside the content class.
This allows for better structure in the html file, which will also
help the page look better in the browser.